debt: Move welcome page to tailwind
