+++ lang/de.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "Du kannst eine zusätzliche Quittung erstellen, die du für steuerlich nicht absetzbare Zahlungen für Dinge wie Veranstaltungstickets, Merchandise oder Dienstleistungen verwenden kannst.",
   "MnpUD7": "Details anzeigen",
   "mNqW2+": "Diese Währung stimmt nicht mit der aus dem Dokument gescannten Währung überein ({value})",
-  "mntSey": "Lege ein Profil für dein Unternehmen an, um als Geldgeber aufzutreten, ermögliche es deinen Mitarbeitern, im Namen deines Unternehmens zu spenden, und vieles mehr.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "z.B. fehlgeschlagener Transfer",
   "MOgYVM": "{amountFrom} bis {amountTo}",
   "MonthlyBudget": "Monatliches Budget",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "Captcha is required.",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "Go to settings",
   "Rqzsq/": "Add Filter",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "A vendor",
