+++ lang/es.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "Puedes crear un recibo adicional para utilizarlo como pago no deducible de impuestos para casos como servicios, mercancía o boletos de eventos.",
   "MnpUD7": "Ver detalles",
   "mNqW2+": "Esta moneda no coincide con la escaneada del documento ({value})",
-  "mntSey": "Crear un perfil para que tu empresa aparezca como colaborador financiero, permitir que tus empleados contribuyan en nombre de tu empresa, y mucho más.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "p. ej., Transferencia fallida",
   "MOgYVM": "{amountFrom} a {amountTo}",
   "MonthlyBudget": "Presupuesto mensual",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "Se requiere CAPTCHA.",
   "rpqkOE": "Especifica el tema o la actividad y el plazo, por ejemplo, \"Formación para voluntarios, abril de 2023\"",
-  "rqE6vX": "Ir a la configuración",
   "Rqzsq/": "Agregar Filtro",
   "RRq5rD": "Esta aplicación puede realizar directamente operaciones críticas que normalmente requerirían 2FA.",
   "rth3eX": "Un proveedor",
