+++ lang/fr.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "Vous pouvez créer un reçu supplémentaire destiné aux paiements non déductibles des impôts, par exemple pour les billets pour les événements, les achats de marchandise ou les services.",
   "MnpUD7": "Voir les détails",
   "mNqW2+": "Cette devise ne correspond pas à celle numérisée à partir du document ({value})",
-  "mntSey": "Créez un profil pour que votre entreprise apparaisse comme contributeur financier, permettez à vos employés de contribuer au nom de votre entreprise, et plus encore.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "Exemple : Transfert échoué",
   "MOgYVM": "{amountFrom} à {amountTo}",
   "MonthlyBudget": "Budget mensuel",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "Captcha requis.",
   "rpqkOE": "Spécifiez l'objet ou l'activité et la période. Par exemple : \"Formation des bénévoles, avril 2023\"",
-  "rqE6vX": "Accéder aux paramètres",
   "Rqzsq/": "Ajouter Filtre",
   "RRq5rD": "Cette application peut effectuer directement des opérations critiques qui nécessitent normalement 2FA.",
   "rth3eX": "Un fournisseur",
