+++ lang/he.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "אפשר ליצור קבלה שאינה על תרומה לשימוש בהוצאה מוכרת (ולא לצרוך זיכוי במס), לתשלומים כגון כרטיסים לאירועים, מכירות או שירותים.",
   "MnpUD7": "View details",
   "mNqW2+": "This currency does not match the one scanned from the document ({value})",
-  "mntSey": "יצירת פרופיל לעסק שלך שיופיע כגורם התומך, ואפשר לעובדים לתרום בשמך או בשם החברה, ועוד.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "e.g. Failed transfer",
   "MOgYVM": "{amountFrom} to {amountTo}",
   "MonthlyBudget": "תקציב חדשי",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "דרוש אימות Captcha.",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "להגדרות",
   "Rqzsq/": "Add Filter",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "A vendor",
