+++ lang/nl.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "You can create an additional receipt for you to use as a non-tax-deductible payments for cases like event tickets, merch, or services.",
   "MnpUD7": "Details weergeven",
   "mNqW2+": "This currency does not match the one scanned from the document ({value})",
-  "mntSey": "Create a profile for your business to appear as financial contributor, enable your employees to contribute on behalf of your company, and more.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "bijv. Mislukte overdracht",
   "MOgYVM": "{amountFrom} tot {amountTo}",
   "MonthlyBudget": "Maandelijks budget",
@@ -3091,7 +3091,6 @@
   "rowNumber": "Nr. {number}",
   "Rpq6pU": "Captcha is vereist.",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "Ga naar instellingen",
   "Rqzsq/": "Filter toevoegen",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "A vendor",
