+++ lang/pl.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "Możesz utworzyć dodatkowy kwit dla siebie, który będzie używany jako płatności bez kosztów uzyskania przychodu w takich przypadkach jak bilety na imprezy, towary lub usługi.",
   "MnpUD7": "View details",
   "mNqW2+": "This currency does not match the one scanned from the document ({value})",
-  "mntSey": "Stwórz profil dla swojej firmy, aby pojawić się jako płatnik, umożliwiając swoim pracownikom dokonywanie wpłat w imieniu firmy i wiele więcej.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "np. Nieudany przelew",
   "MOgYVM": "{amountFrom} to {amountTo}",
   "MonthlyBudget": "Miesięczny budżet",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "Captcha jest wymagana.",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "Przejdź do ustawień",
   "Rqzsq/": "Add Filter",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "A vendor",
