+++ lang/pt-BR.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "You can create an additional receipt for you to use as a non-tax-deductible payments for cases like event tickets, merch, or services.",
   "MnpUD7": "Ver detalhes",
   "mNqW2+": "This currency does not match the one scanned from the document ({value})",
-  "mntSey": "Crie um perfil para que a sua empresa apareça como contribuinte financeiro, permita que seus funcionários contribuam em nome da sua empresa e muito mais.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "e.g. Failed transfer",
   "MOgYVM": "{amountFrom} to {amountTo}",
   "MonthlyBudget": "Monthly budget",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "Captcha is required.",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "Go to settings",
   "Rqzsq/": "Add Filter",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "A vendor",
