+++ lang/sk-SK.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "Môžete si vytvoriť ďalší doklad, ktorý použijete ako nezdaniteľnú platbu v prípadoch, ako sú vstupenky na podujatie, reklamný tovar alebo služby.",
   "MnpUD7": "View details",
   "mNqW2+": "This currency does not match the one scanned from the document ({value})",
-  "mntSey": "Vytvorte profil pre vašu firmu, aby sa zobrazovala ako finančný prispievateľ, umožnite svojim zamestnancom prispievať v mene vašej spoločnosti a podobne.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "e.g. Failed transfer",
   "MOgYVM": "{amountFrom} to {amountTo}",
   "MonthlyBudget": "Mesačný rozpočet",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "Vyžaduje sa Captcha.",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "Prejsť do nastavení",
   "Rqzsq/": "Add Filter",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "A vendor",
