+++ lang/sv-SE.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "Du kan skapa ytterligare ett kvitto för dig att använda som en avdragsgilla betalning som evenemangsbiljetter, varor eller tjänster.",
   "MnpUD7": "View details",
   "mNqW2+": "This currency does not match the one scanned from the document ({value})",
-  "mntSey": "Skapa en profil för att ditt företag ska synas som ekonomisk bidragsgivare. Gör det möjligt för dina medarbetare att bidra på uppdrag av ditt företag, och mer.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "e.g. Failed transfer",
   "MOgYVM": "{amountFrom} till {amountTo}",
   "MonthlyBudget": "Månadsbudget",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "Captcha krävs.",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "Inställningar",
   "Rqzsq/": "Lägg till filter",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "A vendor",
