+++ lang/zh.json
@@ -2427,7 +2427,7 @@
   "MNi3fa": "你可以另外创建收据，让你在活动工单、商品销售或服务等情况下使用非免税付款。",
   "MnpUD7": "View details",
   "mNqW2+": "This currency does not match the one scanned from the document ({value})",
-  "mntSey": "Create a profile for your business to appear as financial contributor, enable your employees to contribute on behalf of your company, and more.",
+  "mntSey": "Create a profile for your business to appear as a financial contributor, enable your employees to contribute on behalf of your company, and more.",
   "mOdpl+": "例如：转账失败",
   "MOgYVM": "{amountFrom} to {amountTo}",
   "MonthlyBudget": "每月预算",
@@ -3091,7 +3091,6 @@
   "rowNumber": "No. {number}",
   "Rpq6pU": "需要进行验证",
   "rpqkOE": "Specify item or activity and timeframe, e.g. \"Volunteer Training, April 2023\"",
-  "rqE6vX": "前往设置",
   "Rqzsq/": "添加筛选",
   "RRq5rD": "This application can directly perform critical operations that would normally require 2FA.",
   "rth3eX": "供应商",
